+++ .storybook-threejs-2d/preview.js
@@ -1,0 +1,1 @@
+export { decorators } from '../src/storybook-channel.js';